metcd must support more things :(
