+++ common/store/metcdstore/metcdstore_test.go
@@ -8,68 +8,10 @@ import (
 	"time"
 
 	"github.com/weaveworks/flux/common/store"
+	"github.com/weaveworks/flux/common/store/test"
 
 	"golang.org/x/net/context"
 )
-
-func TestServiceKeyRegexp(t *testing.T) {
-	for key, want := range map[string]struct {
-		serviceName string
-		ok          bool
-	}{
-		"": {"", false},
-		"/weave-flux/service/foo":           {"", false},
-		"/weave-flux/service/foo/":          {"", false},
-		"/weave-flux/service/foo/details":   {"foo", true},
-		"/weave-flux/service/foo/instance":  {"", false},
-		"/weave-flux/service/foo/groupspec": {"", false},
-	} {
-		serviceName, ok := parseServiceKey([]byte(key))
-		if serviceName != want.serviceName || ok != want.ok {
-			t.Errorf("%q: want (%q %v), have (%q %v)", key, want.serviceName, want.ok, serviceName, ok)
-		}
-	}
-}
-
-func TestInstanceKeyRegexp(t *testing.T) {
-	for key, want := range map[string]struct {
-		serviceName  string
-		instanceName string
-		ok           bool
-	}{
-		"": {"", "", false},
-		"/weave-flux/service/foo":               {"", "", false},
-		"/weave-flux/service/foo/instance/":     {"", "", false},
-		"/weave-flux/service/foo/instance/bar":  {"foo", "bar", true},
-		"/weave-flux/service/foo/instance/bar/": {"", "", false},
-		"/weave-flux/service/foo/groupspec/bar": {"", "", false},
-	} {
-		serviceName, instanceName, ok := parseInstanceKey([]byte(key))
-		if serviceName != want.serviceName || instanceName != want.instanceName || ok != want.ok {
-			t.Errorf("%q: want (%q %q %v), have (%q %q %v)", key, want.serviceName, want.instanceName, want.ok, serviceName, instanceName, ok)
-		}
-	}
-}
-
-func TestContainerRuleKeyRegexp(t *testing.T) {
-	for key, want := range map[string]struct {
-		serviceName       string
-		containerRuleName string
-		ok                bool
-	}{
-		"": {"", "", false},
-		"/weave-flux/service/foo":                {"", "", false},
-		"/weave-flux/service/foo/groupspec/":     {"", "", false},
-		"/weave-flux/service/foo/instance/bar":   {"", "", false},
-		"/weave-flux/service/foo/groupspec/bar":  {"foo", "bar", true},
-		"/weave-flux/service/foo/groupspec/bar/": {"", "", false},
-	} {
-		serviceName, containerRuleName, ok := parseContainerRuleKey([]byte(key))
-		if serviceName != want.serviceName || containerRuleName != want.containerRuleName || ok != want.ok {
-			t.Errorf("%q: want (%q %q %v), have (%q %q %v)", key, want.serviceName, want.containerRuleName, want.ok, serviceName, containerRuleName, ok)
-		}
-	}
-}
 
 func TestMetcdStore(t *testing.T) {
 	// This is a stopgap solution until metcdstore implements the complete Store
@@ -208,5 +150,20 @@ func TestMetcdStore(t *testing.T) {
 	}
 	if want, have := 0, len(serviceInfos); want != have {
 		t.Fatalf("GetAllServices: want %d, have %d", want, have)
+	}
+}
+
+func DontTestMetcdStoreSuite(t *testing.T) {
+	logger := log.New(ioutil.Discard, "", log.LstdFlags)
+	s := New(context.Background(), 1, logger)
+	time.Sleep(3 * time.Second)
+	test.RunStoreTestSuite(testableStore{s}, t)
+}
+
+type testableStore struct{ store.Store }
+
+func (ts testableStore) Reset(t *testing.T) {
+	if err := ts.Store.RemoveAllServices(); err != nil {
+		t.Fatalf("during Reset: %v", err)
 	}
 }
